Rename incremental generator files
